graphql: change audio duration from int to float
Summary: Fixes https://app.asana.com/0/75307071900036/82088038130698

Reviewers: kunaljham, michaelraines

Reviewed By: michaelraines

Subscribers: janky

Differential Revision: https://phabricator.corp.sprucehealth.com/D389

+++ cmd/svc/baymaxgraphql/graphql_thread_item.go
@@ -199,7 +199,7 @@ var audioAttachmentType = graphql.NewObject(
 		Fields: graphql.Fields{
 			"mimetype":          &graphql.Field{Type: graphql.NewNonNull(graphql.String)},
 			"url":               &graphql.Field{Type: graphql.NewNonNull(graphql.String)},
-			"durationInSeconds": &graphql.Field{Type: graphql.NewNonNull(graphql.Int)},
+			"durationInSeconds": &graphql.Field{Type: graphql.NewNonNull(graphql.Float)},
 		},
 		IsTypeOf: func(value interface{}, info graphql.ResolveInfo) bool {
 			_, ok := value.(*audioAttachment)
